Refactor to be a class
