--- packages/jbpm-form-generator/tests/getFormGenerator.test.ts
@@ -1,51 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { getFormGenerator, registerFormGeneratorType } from "../dist/getFormGenerator";
-import { FormGenerator, FormStyle } from "../dist/types";
-
-describe("getFormGenerator tests", () => {
-  it("Lookup existing tool - patternfly", () => {
-    const tool = getFormGenerator(FormStyle.PATTERNFLY);
-
-    expect(tool).not.toBeUndefined();
-    expect(tool.type).toStrictEqual(FormStyle.PATTERNFLY);
-  });
-
-  it("Lookup wrong tool", () => {
-    const toolType = "wrong tool type";
-    expect(() => getFormGenerator(toolType)).toThrow(`Unsupported form generation type: "${toolType}"`);
-  });
-
-  it("Register tool & lookup", () => {
-    const tool: FormGenerator = {
-      type: "cool new tool",
-      generate: jest.fn(),
-    };
-
-    registerFormGeneratorType(tool);
-
-    const coolTool = getFormGenerator(tool.type);
-    expect(coolTool).not.toBeUndefined();
-    expect(coolTool).toStrictEqual(tool);
-
-    const patternfly = getFormGenerator(FormStyle.PATTERNFLY);
-    expect(patternfly).not.toBeUndefined();
-  });
-});
